==> Sergey/0_Crime_Data_Initial_Merge.ipynb <==
import pandas as pd
import os

crime_2020_df = pd.read_excel("data/NIBRSPublicView.Jan1-Dec31-2020.xlsx")


crime_2021_df = pd.read_excel("data/NIBRSPublicViewSep21.xlsx")

#Clean up the headers 2020
crime_2020_df = crime_2020_df.rename(columns = {  "Occurrence\nDate": "Date", 
                                                  "Occurrence\nHour": "Hour",
                                                  "NIBRS\nClass": "NIBRS_Class",
                                                  "NIBRSDescription": "Offense_Type",
                                                  "Offense\nCount": "Offense_Count",
                                                  "Block Range": "Street_No",
                                                  "StreetName": "Street_Name",
                                                  "Street\nType": "Street_Type",
                                                  "ZIP Code": "Zip_Code"
                                               }
                                    )
    
crime_2020_df.head() 

crime_2020_df

#Clean up the headers 2021
crime_2021_df = crime_2021_df.rename(columns = {  "RMSOccurrenceDate": "Date", 
                                                  "RMSOccurrenceHour": "Hour",
                                                  "NIBRSClass": "NIBRS_Class",
                                                  "NIBRSDescription": "Offense_Type",
                                                  "OffenseCount": "Offense_Count",
                                                  "StreetNo": "Street_No",
                                                  "StreetName": "Street_Name",
                                                  "StreetType": "Street_Type",
                                                  "ZIPCode": "Zip_Code"
                                               }
                                    )
    
crime_2021_df.head() 

crime_2021_df

#Check Headers
crime_2020_df.columns

#Check Headers
crime_2021_df.columns

#Concatenate data frames 
data_frames = [crime_2020_df, crime_2021_df]
crime_df = pd.concat(data_frames, sort=False)
crime_df

#Makes an output directory
dir = r"Output"
if not os.path.exists(dir):
    os.mkdir(dir)

crime_df.to_csv('Output/1_crime_data_merged.csv', index=False)
