==> Sergey/3_Appending_with_Lat_Long_data.ipynb <==
import pandas as pd
import os

#import lat long output 0-150000
lat_long =["Output/Lat_Long_0_150000.csv",
          "Output/Lat_Long_150000_354136.csv"]

temp_list = []

for csv in lat_long:
    temp_df = pd.read_csv(csv, index_col=None, header=0)
    temp_list.append(temp_df)

full_lat_long = pd.concat(temp_list, axis=0, ignore_index=True)

full_lat_long

crime_df = pd.read_csv("Output/2_crime_data_initial_clean.csv")
crime_df

#Append master list with lat long data
crime_df['Latitude'] = full_lat_lngs['Latitude']
crime_df['Longitude'] = full_lat_lngs['Longitude']
crime_df

crime_df.to_csv('Output/3_crime_data_lat_long.csv', index=False,)
