==> Sergey/5_ZipCode_Scraper_Update.ipynb <==
# Dependencies
from bs4 import BeautifulSoup as bs
import requests
from splinter import Browser
import os
import pandas as pd 

#Set up for chromedriver.exe
executable_path = {"executable_path": '/Users/Serj/Documents/Chromedriver/chromedriver.exe'}
browser = Browser('chrome', **executable_path, headless=False)

#Establish a connection for the nasa website
url = 'https://www.unitedstateszipcodes.org/'


crime_cleaned_df = pd.read_csv('Output/3_crime_data_lat_long.csv')
crime_cleaned_df

zip_list = crime_cleaned_df.Zip_Code.unique().tolist()
zip_list

new_url = "https://www.unitedstateszipcodes.org/77062/"
browser.visit(new_url)
zip_browse = browser.html
zip_soup = bs(zip_browse,'lxml')
data_parse = zip_soup.find_all("table",class_ = "table table-hover")

data_parse = zip_soup.find_all("table",class_ = "table table-hover")
data_parse2 = zip_soup.find_all("table",class_ = "chart-legend table table-striped table-hover table-condensed")
data = {}

 #################################################################################
#Population Information
tds = data_parse[0].find_all("td",class_ = "text-right")
data["Population"] = tds[0].text
data["Population Density"] = tds[1].text
data["Housing Units"] = tds[2].text
data["Median Home Value"] = tds[3].text
tds = data_parse[1].find_all("td",class_ = "text-right")
data["Occupied_Housing_Units"] = tds[2].text
data["Median_Household_Income"] = tds[3].text

#################################################################################
#Gender
tds = data_parse2[0].find_all("span",class_ = "value")
data["Gender_Male"] = tds[0].text
data["Gender_Female"] = tds[1].text

#################################################################################
#Race
tds = data_parse2[1].find_all("span",class_ = "value")
data["White"] = tds[0].text
data["Black"] = tds[1].text
data["Indian"] = tds[2].text
data["Hawaiian"] = tds[3].text
data["Other"] = tds[4].text
data["Mixed"] = tds[5].text

#################################################################################
#Housing arrangement
tds = data_parse2[2].find_all("span",class_ = "value")
data["Whole_Family"] = tds[0].text
data["Single_Guardian"] = tds[1].text
data["Singles"] = tds[2].text
data["Single_With_Roomate"] = tds[3].text
#################################################################################
#Renter vs Mortgage
tds = data_parse2[5].find_all("span",class_ = "value")
data["Owned_With_Mortgage"] = tds[0].text
data["Owned_No_Mortgage"] = tds[1].text
data["Renter"] = tds[2].text
data["Vacant"] = tds[3].text

#################################################################################
#Employment
tds = data_parse2[8].find_all("span",class_ = "value")
data["Full_Time"] = tds[0].text
data["Part_Time"] = tds[1].text
data["Unemployed"] = tds[2].text


data

executable_path = {"executable_path": '/Users/Serj/Documents/Chromedriver/chromedriver.exe'}
browser = Browser('chrome', **executable_path, headless=False)
url = 'https://www.unitedstateszipcodes.org/'

from random import randint
from time import sleep


population_data = [] 



for x in zip_list:
    
    try:
        sleep(randint(5,15))
        #Create an empty dictionary
        zip_dict = {}
        #Add the zipcode on the end of the url
        new_url = f"{url}{x}/"
        print(new_url)

        browser.visit(new_url)
        zip_browse = browser.html

        zip_soup = bs(zip_browse,'lxml')

        data_parse = zip_soup.find_all("div",class_ = "col-xs-12 col-sm-6")
        data_parse = zip_soup.find_all("table",class_ = "table table-hover")
        data = {}

        data_parse = zip_soup.find_all("table",class_ = "table table-hover")
        data_parse2 = zip_soup.find_all("table",class_ = "chart-legend table table-striped table-hover table-condensed")
        data = {}


        #################################################################################
        #ZipCode     
        data['Zipcode'] = x
        #################################################################################
        #Population Information
        tds = data_parse[0].find_all("td",class_ = "text-right")
        data["Population"] = tds[0].text
        data["Population Density"] = tds[1].text
        data["Housing Units"] = tds[2].text
        data["Median Home Value"] = tds[3].text
        tds = data_parse[1].find_all("td",class_ = "text-right")
        data["Occupied_Housing_Units"] = tds[2].text
        data["Median_Household_Income"] = tds[3].text

        #################################################################################
        #Gender
        tds = data_parse2[0].find_all("span",class_ = "value")
        data["Gender_Male"] = tds[0].text
        data["Gender_Female"] = tds[1].text

        #################################################################################
        #Race
        tds = data_parse2[1].find_all("span",class_ = "value")
        data["White"] = tds[0].text
        data["Black"] = tds[1].text
        data["Indian"] = tds[2].text
        data["Asian"] = tds[3].text
        data["Hawaiian"] = tds[4].text
        data["Other"] = tds[5].text
        data["Mixed"] = tds[6].text

        #################################################################################
        #Housing arrangement
        tds = data_parse2[2].find_all("span",class_ = "value")
        data["Whole_Family"] = tds[0].text
        data["Single_Guardian"] = tds[1].text
        data["Singles"] = tds[2].text
        data["Single_With_Roomate"] = tds[3].text
        #################################################################################
        #Renter vs Mortgage
        tds = data_parse2[5].find_all("span",class_ = "value")
        data["Owned_With_Mortgage"] = tds[0].text
        data["Owned_No_Mortgage"] = tds[1].text
        data["Renter"] = tds[2].text
        data["Vacant"] = tds[3].text

        #################################################################################
        #Employment
        tds = data_parse2[8].find_all("span",class_ = "value")
        data["Full_Time"] = tds[0].text
        data["Part_Time"] = tds[1].text
        data["Unemployed"] = tds[2].text

        population_data.append(data)

        print("-----------")
        print(x)
        print("done")
        print("----------")
    except IndexError:
        print("Sorry Can't find value for " )
        print(x)
        print("----------------------------")

pd.DataFrame(population_data)

df = pd.DataFrame(population_data)
df

df.to_csv('Output/Population_Data.csv', index=False,)
