==> Sergey/6_Appending_with_Tract_data.ipynb <==
import pandas as pd
import os

crime_df = pd.read_csv("Output/3_crime_data_lat_long.csv")
crime_df

tract =["Output/Tract_0_50000.csv",
        "Output/Tract_50000_100000.csv",
        "Output/Tract_100000_150000.csv",
        "Output/Tract_150000_200000.csv",
        "Output/Tract_200000_250000.csv",
        "Output/Tract_250000_300000.csv",
        "Output/Tract_300000_354135.csv"]

temp_list = []

for csv in tract:
    temp_df = pd.read_csv(csv, index_col=None, header=0)
    temp_list.append(temp_df)

tract_df = pd.concat(temp_list, axis=0, ignore_index=True)

tract_df

crime_df['Census_Tract'] = tract_df['Census_Tract']
crime_df

crime_df.to_csv('Output/4_crime_data_lat_long_tract.csv', index=False,)
