==> Sergey/7_Remove_Blank_LatLong_Tract.ipynb <==
import pandas as pd
import os

crime_df = pd.read_csv("Output/4_crime_data_lat_long_tract.csv")
crime_df

crime_df.dropna(subset = ['Latitude'], inplace = True) 
crime_df

crime_df.to_csv('Output/6_Cleaned_lat_long.csv', index=False,)
